Add initial representations for decision variables
