+++ crates/shackle-runtime/src/value.rs
@@ -1,10 +1,15 @@
+mod bool_var;
+mod float_var;
+mod int_set_var;
+mod int_var;
+mod num;
+mod seq;
+mod set;
+
 use std::{
-	collections::{HashMap, VecDeque},
 	fmt::Debug,
-	iter::once,
 	mem::MaybeUninit,
 	num::NonZeroU64,
-	ops::RangeInclusive,
 	pin::Pin,
 	ptr::NonNull,
 	str::{from_utf8, from_utf8_unchecked},
@@ -19,29 +24,26 @@ use std::{
 
 use bilge::{
 	bitsize,
-	prelude::{u10, u2, u52, u61, Number},
+	prelude::{u2, u61, Number},
 	Bitsized, TryFromBits,
 };
-use itertools::Itertools;
 use once_cell::sync::Lazy;
+pub use set::SetView;
 use varlen::{
-	array_init::MoveFrom,
 	define_varlen,
 	prelude::{ArrayInitializer, FromIterPrefix},
 	Initializer, Layout, VarLen,
 };
 
-use self::{
+use crate::value::{
+	bool_var::{BoolVar, BoolVarRef},
+	float_var::{FloatVar, FloatVarRef},
+	int_set_var::{IntSetVar, IntSetVarRef},
+	int_var::{IntVar, IntVarRef},
 	num::{FloatVal, IntVal},
 	seq::{InnerViewType, Pairs, SeqView, ViewType},
 	set::{FloatSetView, IntSetView},
 };
-
-mod num;
-mod seq;
-mod set;
-
-pub use set::SetView;
 
 #[bitsize(2)]
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, TryFromBits)]
@@ -231,10 +233,16 @@ impl Value {
 							},
 						})
 					}
-					ValType::BoolVar => todo!(),
-					ValType::IntVar => todo!(),
-					ValType::FloatVar => todo!(),
-					ValType::IntSetVar => todo!(),
+					ValType::BoolVar => {
+						DataView::BoolVar(v.refs().bool_var[0].lock().unwrap().into())
+					}
+					ValType::IntVar => DataView::IntVar(v.refs().int_var[0].lock().unwrap().into()),
+					ValType::FloatVar => {
+						DataView::FloatVar(v.refs().float_var[0].lock().unwrap().into())
+					}
+					ValType::IntSetVar => {
+						DataView::IntSetVar(v.refs().int_set_var[0].lock().unwrap().into())
+					}
 				}
 			},
 		}
@@ -264,224 +272,6 @@ impl Value {
 		}
 	}
 
-	fn new_boxed_int(i: IntVal) -> Value {
-		let (b, i) = match i {
-			IntVal::InfPos => (true, 1),
-			IntVal::InfNeg => (true, -1),
-			IntVal::Int(i) => (false, i),
-		};
-		Self::new_box(value_storage::Init {
-			ty: ValType::Int,
-			len: 0,
-			ref_count: 1.into(),
-			weak_count: 0.into(),
-			values: InitEmpty,
-			ints: FromIterPrefix([i].into_iter()),
-			floats: InitEmpty,
-			bytes: FromIterPrefix([b as u8].into_iter()),
-		})
-	}
-
-	fn new_boxed_float(i: FloatVal) -> Value {
-		Self::new_box(value_storage::Init {
-			ty: ValType::Float,
-			len: 0,
-			ref_count: 1.into(),
-			weak_count: 0.into(),
-			values: InitEmpty,
-			ints: InitEmpty,
-			floats: FromIterPrefix([i].into_iter()),
-			bytes: InitEmpty,
-		})
-	}
-
-	/// Create a sequence view with custom index set dimensions
-	///
-	/// # Warning
-	/// This method will panic if it is called on a non-sequence value, or if the
-	/// size of the provided dimensions do not equal the number of elements in the
-	/// sequence.
-	pub fn with_dim<D: IntoIterator<Item = (i64, i64)>>(&self, dims: D) -> Value {
-		assert!(matches!(self.deref(), DataView::Seq(_)));
-		let DataView::Seq(seq) = self.deref() else {
-			panic!("unable to give dimensions to non-sequence value");
-		};
-		let dims = dims.into_iter().flat_map(|(a, b)| [a, b]).collect_vec();
-		assert!(
-			seq.len()
-				== dims
-					.iter()
-					.tuples()
-					.map(|(a, b)| (b - a + 1) as usize)
-					.product::<usize>(),
-			"dimensions given do not match the length of the array"
-		);
-		match seq {
-			SeqView::Slice {
-				dims: _,
-				slice,
-				storage: _,
-			} => Self::new_box(value_storage::Init {
-				ty: ValType::View,
-				ref_count: 1.into(),
-				weak_count: 0.into(),
-				len: ViewType::new(
-					InnerViewType::Slice,
-					(dims.len() / 2) as u8,
-					slice.len() as u8,
-				)
-				.as_len(),
-				values: MoveFrom([self.clone()]),
-				ints: FromIterPrefix(
-					dims.into_iter()
-						.chain(slice.iter().flat_map(|(&min, &max)| [min, max])),
-				),
-				floats: InitEmpty,
-				bytes: InitEmpty,
-			}),
-			SeqView::Compressed {
-				dims: _,
-				repeat,
-				storage: _,
-			} => Self::new_box(value_storage::Init {
-				ty: ValType::View,
-				ref_count: 1.into(),
-				weak_count: 0.into(),
-				len: ViewType::new(InnerViewType::Compact, (dims.len() / 2) as u8, 0).as_len(),
-				values: MoveFrom([self.clone()]),
-				ints: FromIterPrefix(once(repeat).chain(dims)),
-				floats: InitEmpty,
-				bytes: InitEmpty,
-			}),
-			SeqView::WithDim { dims: _, storage } => Self::new_box(value_storage::Init {
-				ty: ValType::View,
-				ref_count: 1.into(),
-				weak_count: 0.into(),
-				len: ViewType::new(InnerViewType::Dim, (dims.len() / 2) as u8, 0).as_len(),
-				values: MoveFrom([storage.clone()]),
-				ints: FromIterPrefix(dims.into_iter()),
-				floats: InitEmpty,
-				bytes: InitEmpty,
-			}),
-			_ => Self::new_box(value_storage::Init {
-				ty: ValType::View,
-				ref_count: 1.into(),
-				weak_count: 0.into(),
-				len: ViewType::new(InnerViewType::Dim, (dims.len() / 2) as u8, 0).as_len(),
-				values: MoveFrom([self.clone()]),
-				ints: FromIterPrefix(dims.into_iter()),
-				floats: InitEmpty,
-				bytes: InitEmpty,
-			}),
-		}
-	}
-
-	/// Create a slice view of a sequence
-	///
-	/// This creates a view that occludes part of the underlying sequence, and
-	/// optionally gives the view new dimensions.
-	///
-	/// # Warning
-	/// This method will panic if the underlying value is a non-sequence value, if
-	///  the sequence is sliced outside its underlying index set(s), or if the
-	///  number of non-occluded elements does not equal the size of the provided
-	///  dimensions
-	pub fn slice<
-		It1: ExactSizeIterator<Item = (i64, i64)>,
-		It2: ExactSizeIterator<Item = (i64, i64)>,
-		I: IntoIterator<IntoIter = It1>,
-		J: IntoIterator<IntoIter = It2>,
-	>(
-		&self,
-		select_idxs: J,
-		view_dims: I,
-	) -> Value {
-		let DataView::Seq(seq) = self.deref() else {
-			panic!("unable to give dimensions to non-sequence value");
-		};
-		let slice: Vec<i64> = select_idxs
-			.into_iter()
-			.flat_map(|(start, end)| [start, end])
-			.collect();
-		assert_eq!(
-			seq.dims(),
-			slice.len() / 2,
-			"unable to slice a sequence with {} dimensions, using {} sets",
-			seq.dims(),
-			slice.len() / 2
-		);
-
-		assert!(
-			slice
-				.iter()
-				.tuples()
-				.zip(1..=seq.dims())
-				.all(|((start, end), d)| {
-					let (d_start, d_end) = seq.dim(d);
-					d_start <= *start && *end <= d_end
-				}),
-			"slicing index out of bounds"
-		);
-
-		let dims: Vec<i64> = view_dims
-			.into_iter()
-			.flat_map(|(start, end)| [start, end])
-			.collect();
-		assert_eq!(
-			dims.iter().tuples().map(|(start, end)| end - start + 1).product::<i64>(),
-			slice.iter().tuples().map(|(start, end)| end - start + 1).product::<i64>(),
-			"size of the dimensions provided for the slice does not match the number of elements in sliced sequence"
-		);
-		// TODO: See what underlying view could be incorporated in the slice. (WithDim and Slice?)
-		Self::new_box(value_storage::Init {
-			ty: ValType::View,
-			ref_count: 1.into(),
-			weak_count: 0.into(),
-			len: ViewType::new(
-				InnerViewType::Slice,
-				(dims.len() / 2) as u8,
-				(slice.len() / 2) as u8,
-			)
-			.as_len(),
-			values: MoveFrom([self.clone()]),
-			ints: FromIterPrefix(dims.into_iter().chain(slice)),
-			floats: InitEmpty,
-			bytes: InitEmpty,
-		})
-	}
-
-	/// Create a sequence view transposing an existing view
-	///
-	/// The arguments of this method are the number of the index sets to which the
-	/// n-th index will be translated. Negative numbers can be used to reverse a
-	/// dimension.
-	///
-	/// # Warning
-	/// This method will panic if it is called on a non-sequence value, or if it
-	/// is provided with a number for an index set that is beyond the possible
-	/// number of index sets.
-	pub fn transpose<D: IntoIterator<Item = i64>>(&self, dims: D) -> Value {
-		let DataView::Seq(seq) = self.deref() else {
-			panic!("unable to give dimensions to non-sequence value");
-		};
-		let dims = dims.into_iter().collect_vec();
-		assert!(
-			dims.iter()
-				.all(|i| i.unsigned_abs() as usize <= seq.dims() && *i != 0),
-			"invalid index set reference"
-		);
-		Self::new_box(value_storage::Init {
-			ty: ValType::View,
-			ref_count: 1.into(),
-			weak_count: 0.into(),
-			len: ViewType::new(InnerViewType::Transpose, dims.len() as u8, 0).as_len(),
-			values: MoveFrom([self.clone()]),
-			ints: FromIterPrefix(dims.into_iter()),
-			floats: InitEmpty,
-			bytes: InitEmpty,
-		})
-	}
-
 	pub fn new_str<I: ExactSizeIterator<Item = u8>, S: IntoIterator<IntoIter = I>>(s: S) -> Value {
 		let s = s.into_iter();
 		if s.len() == 0 {
@@ -495,6 +285,10 @@ impl Value {
 			values: InitEmpty,
 			ints: InitEmpty,
 			floats: InitEmpty,
+			bool_var: InitEmpty,
+			int_var: InitEmpty,
+			float_var: InitEmpty,
+			int_set_var: InitEmpty,
 			bytes: FromIterPrefix(s),
 		});
 		debug_assert!(from_utf8(v.get_pin().refs().bytes).is_ok());
@@ -521,248 +315,13 @@ pub static EMPTY_STRING: Lazy<Value> = Lazy::new(|| {
 		values: InitEmpty,
 		ints: InitEmpty,
 		floats: InitEmpty,
+		bool_var: InitEmpty,
+		int_var: InitEmpty,
+		float_var: InitEmpty,
+		int_set_var: InitEmpty,
 		bytes: InitEmpty,
 	})
 });
-
-pub static EMPTY_SEQ: Lazy<Value> = Lazy::new(|| {
-	Value::new_box(value_storage::Init {
-		ty: ValType::Seq,
-		len: 0u32,
-		ref_count: 1.into(),
-		weak_count: 0.into(),
-		values: InitEmpty,
-		ints: InitEmpty,
-		floats: InitEmpty,
-		bytes: InitEmpty,
-	})
-});
-
-static INT_MAP: Lazy<Mutex<HashMap<IntVal, Value>>> = Lazy::new(|| HashMap::new().into());
-pub static INT_INF_POS: Lazy<Value> = Lazy::new(|| {
-	let mut map = INT_MAP.lock().unwrap();
-	let inf = map
-		.entry(IntVal::InfPos)
-		.or_insert_with(|| Value::new_boxed_int(IntVal::InfPos));
-	inf.clone()
-});
-pub static INT_INF_NEG: Lazy<Value> = Lazy::new(|| {
-	let mut map = INT_MAP.lock().unwrap();
-	let inf = map
-		.entry(IntVal::InfNeg)
-		.or_insert_with(|| Value::new_boxed_int(IntVal::InfNeg));
-	inf.clone()
-});
-
-static FLOAT_MAP: Lazy<Mutex<HashMap<FloatVal, Value>>> = Lazy::new(|| HashMap::new().into());
-#[allow(dead_code)] // TODO!
-pub static FLOAT_INF_POS: Lazy<Value> = Lazy::new(|| FloatVal::INFINITY.into());
-#[allow(dead_code)] // TODO!
-pub static FLOAT_INF_NEG: Lazy<Value> = Lazy::new(|| FloatVal::NEG_INFINITY.into());
-
-pub static INT_SET_EMPTY: Lazy<Value> = Lazy::new(|| {
-	Value::new_box(value_storage::Init {
-		ty: ValType::IntSet,
-		len: 0b11,
-		ref_count: 1.into(),
-		weak_count: 0.into(),
-		values: InitEmpty,
-		ints: FromIterPrefix([1i64, 0].into_iter()),
-		floats: InitEmpty,
-		bytes: InitEmpty,
-	})
-});
-pub static INT_SET_INF: Lazy<Value> = Lazy::new(|| {
-	Value::new_box(value_storage::Init {
-		ty: ValType::IntSet,
-		len: 0b00,
-		ref_count: 1.into(),
-		weak_count: 0.into(),
-		values: InitEmpty,
-		ints: InitEmpty,
-		floats: InitEmpty,
-		bytes: InitEmpty,
-	})
-});
-
-pub static FLOAT_SET_EMPTY: Lazy<Value> = Lazy::new(|| {
-	Value::new_box(value_storage::Init {
-		ty: ValType::FloatSet,
-		len: 0,
-		ref_count: 1.into(),
-		weak_count: 0.into(),
-		values: InitEmpty,
-		ints: InitEmpty,
-		floats: InitEmpty,
-		bytes: InitEmpty,
-	})
-});
-pub static FLOAT_SET_INF: Lazy<Value> = Lazy::new(|| {
-	Value::new_box(value_storage::Init {
-		ty: ValType::FloatSet,
-		len: 1,
-		ref_count: 1.into(),
-		weak_count: 0.into(),
-		values: InitEmpty,
-		ints: InitEmpty,
-		floats: FromIterPrefix([FloatVal::NEG_INFINITY, FloatVal::INFINITY].into_iter()),
-		bytes: InitEmpty,
-	})
-});
-
-impl From<bool> for Value {
-	fn from(value: bool) -> Self {
-		Value::from(if value { 1 } else { 0 })
-	}
-}
-impl TryInto<bool> for &Value {
-	type Error = ();
-
-	fn try_into(self) -> Result<bool, Self::Error> {
-		let val: i64 = self.try_into()?;
-		if val != 0 && val != 1 {
-			todo!()
-		}
-		Ok(val == 1)
-	}
-}
-impl TryInto<bool> for Value {
-	type Error = ();
-	fn try_into(self) -> Result<bool, Self::Error> {
-		(&self).try_into()
-	}
-}
-
-impl From<IntVal> for Value {
-	fn from(value: IntVal) -> Self {
-		match value {
-			IntVal::InfPos => INT_INF_POS.clone(),
-			IntVal::InfNeg => INT_INF_NEG.clone(),
-			IntVal::Int(i) if (Self::MIN_INT..=Self::MAX_INT).contains(&i) => {
-				// Can box integer (fits in 62 bits)
-				let mut x = i.unsigned_abs() << 3;
-				if i.is_negative() {
-					x |= Self::INT_SIGN_BIT;
-				}
-				x |= Self::INT_TAG;
-				Self {
-					raw: NonZeroU64::new(x).unwrap(),
-				}
-			}
-			iv => {
-				// Try and find integer in map or allocate new integer
-				let mut map = INT_MAP.lock().unwrap();
-				let v = map.entry(iv).or_insert_with(|| Value::new_boxed_int(iv));
-				v.clone()
-			}
-		}
-	}
-}
-impl TryInto<IntVal> for &Value {
-	type Error = ();
-	fn try_into(self) -> Result<IntVal, Self::Error> {
-		if let DataView::Int(i) = self.deref() {
-			Ok(i)
-		} else {
-			todo!()
-		}
-	}
-}
-impl TryInto<IntVal> for Value {
-	type Error = ();
-	fn try_into(self) -> Result<IntVal, Self::Error> {
-		(&self).try_into()
-	}
-}
-impl TryInto<i64> for &Value {
-	type Error = ();
-	fn try_into(self) -> Result<i64, Self::Error> {
-		if let IntVal::Int(i) = self.try_into()? {
-			Ok(i)
-		} else {
-			Err(())
-		}
-	}
-}
-impl TryInto<i64> for Value {
-	type Error = ();
-	fn try_into(self) -> Result<i64, Self::Error> {
-		(&self).try_into()
-	}
-}
-impl From<i64> for Value {
-	fn from(value: i64) -> Self {
-		IntVal::Int(value).into()
-	}
-}
-
-impl From<FloatVal> for Value {
-	fn from(value: FloatVal) -> Self {
-		let f: f64 = value.into();
-		const EXPONENT_MASK: u64 = 0x7FF << 52;
-		let bits = f.to_bits();
-		let mut exponent = (bits & EXPONENT_MASK) >> 52;
-		if exponent != 0 {
-			if !(513..=1534).contains(&exponent) {
-				// Exponent doesn't fit in 10 bits
-				let mut map = FLOAT_MAP.lock().unwrap();
-				let v = map
-					.entry(value)
-					.or_insert_with(|| Value::new_boxed_float(value));
-				return v.clone();
-			}
-			exponent -= 512; // Make exponent fit in 10 bits, with bias 511
-		}
-		debug_assert!(exponent <= <u10 as Number>::MAX.value().into());
-		let sign = (bits & (1 << 63)) != 0;
-
-		const FRACTION_MASK: u64 = 0xFFFFFFFFFFFFF;
-		let fraction = bits & FRACTION_MASK; // Remove one bit of precision
-		debug_assert!(fraction <= <u52 as Number>::MAX.value());
-		let mut raw = (fraction << 1) | (exponent << 53) | Self::FLOAT_TAG;
-		if sign {
-			raw |= Self::FLOAT_SIGN_BIT;
-		}
-		Value {
-			raw: NonZeroU64::new(raw).unwrap(),
-		}
-	}
-}
-impl From<f64> for Value {
-	fn from(value: f64) -> Self {
-		Value::from(FloatVal::from(value))
-	}
-}
-impl TryInto<FloatVal> for &Value {
-	type Error = ();
-	fn try_into(self) -> Result<FloatVal, Self::Error> {
-		if let DataView::Float(f) = self.deref() {
-			Ok(f)
-		} else {
-			todo!()
-		}
-	}
-}
-impl TryInto<FloatVal> for Value {
-	type Error = ();
-	fn try_into(self) -> Result<FloatVal, Self::Error> {
-		(&self).try_into()
-	}
-}
-impl TryInto<f64> for &Value {
-	type Error = ();
-
-	fn try_into(self) -> Result<f64, Self::Error> {
-		let fv: FloatVal = self.try_into()?;
-		Ok(fv.into())
-	}
-}
-impl TryInto<f64> for Value {
-	type Error = ();
-	fn try_into(self) -> Result<f64, Self::Error> {
-		(&self).try_into()
-	}
-}
 
 impl From<&str> for Value {
 	fn from(value: &str) -> Self {
@@ -781,140 +340,7 @@ impl<'a> TryInto<&'a str> for &'a Value {
 	}
 }
 
-impl FromIterator<Value> for Value {
-	fn from_iter<T: IntoIterator<Item = Value>>(iter: T) -> Self {
-		let v = iter.into_iter().collect_vec();
-		if v.is_empty() {
-			return EMPTY_SEQ.clone();
-		}
-		let mut eq_count = 0;
-		if v.len() >= 4 {
-			eq_count = v.iter().skip(1).take_while(|&x| x == &v[0]).count();
-		}
-		let iter = v.into_iter().skip(if eq_count >= 3 { eq_count } else { 0 });
-
-		let len = iter.len() as u32;
-		let mut val = Self::new_box(value_storage::Init {
-			ty: ValType::Seq,
-			ref_count: 1.into(),
-			weak_count: 0.into(),
-			len,
-			values: FromIterPrefix(iter),
-			ints: InitEmpty,
-			floats: InitEmpty,
-			bytes: InitEmpty,
-		});
-
-		if eq_count >= 3 {
-			val = Self::new_box(value_storage::Init {
-				ty: ValType::View,
-				ref_count: 1.into(),
-				weak_count: 0.into(),
-				len: ViewType::new(InnerViewType::Compact, 1, 0).as_len(),
-				values: MoveFrom([val]),
-				ints: MoveFrom([eq_count as i64, 1, len as i64]),
-				floats: InitEmpty,
-				bytes: InitEmpty,
-			});
-		}
-		val
-	}
-}
-impl FromIterator<RangeInclusive<IntVal>> for Value {
-	fn from_iter<T: IntoIterator<Item = RangeInclusive<IntVal>>>(iter: T) -> Self {
-		let mut values: VecDeque<_> = iter
-			.into_iter()
-			.filter(|r| r.start() <= r.end())
-			.coalesce(|r1, r2| match (r1.end(), r2.start()) {
-				(IntVal::Int(i), IntVal::Int(j)) if i + 1 >= *j => Ok(*r1.start()..=*r2.end()),
-				(a, b) if a >= b => Ok(*r1.start()..=*r2.end()),
-				_ => Err((r1, r2)),
-			})
-			.flat_map(|r| [*r.start(), *r.end()].into_iter())
-			.collect();
-
-		// Only create a single empty / infinity set
-		if values.is_empty() {
-			return INT_SET_EMPTY.clone();
-		} else if matches!(
-			&values.as_slices(),
-			(&[IntVal::InfNeg, IntVal::InfPos], &[])
-		) {
-			return INT_SET_INF.clone();
-		}
-
-		// Number of counted intervals (for storage lb/ub are counted using flags)
-		let len = values.len() / 2 - 1;
-		assert!(len < 2_usize.pow(31));
-		let mut len = (len << 2) as u32;
-		if matches!(values.front().unwrap(), IntVal::Int(_)) {
-			len |= 0b01;
-		} else {
-			values.pop_front();
-		}
-		if matches!(values.back().unwrap(), IntVal::Int(_)) {
-			len |= 0b10;
-		} else {
-			values.pop_back();
-		}
-
-		debug_assert_eq!(ValueStorage::int_set_len(len), values.len());
-
-		Self::new_box(value_storage::Init {
-			ty: ValType::IntSet,
-			len,
-			ref_count: 1.into(),
-			weak_count: 0.into(),
-			values: InitEmpty,
-			ints: FromIterPrefix(values.into_iter().map(|i| {
-				let IntVal::Int(i) = i else { unreachable!() };
-				i
-			})),
-			floats: InitEmpty,
-			bytes: InitEmpty,
-		})
-	}
-}
-
-impl FromIterator<RangeInclusive<FloatVal>> for Value {
-	fn from_iter<T: IntoIterator<Item = RangeInclusive<FloatVal>>>(iter: T) -> Self {
-		let values: Vec<FloatVal> = iter
-			.into_iter()
-			.filter(|r| r.start() <= r.end())
-			.coalesce(|r1, r2| {
-				if r1.end() >= r2.start() {
-					Ok(*r1.start()..=*r2.end())
-				} else {
-					Err((r1, r2))
-				}
-			})
-			.flat_map(|r| [*r.start(), *r.end()].into_iter())
-			.collect();
-
-		// Use only single empty and (full) infinity set
-		if values.is_empty() {
-			return FLOAT_SET_EMPTY.clone();
-		} else if values.len() == 2
-			&& values[0] == FloatVal::NEG_INFINITY
-			&& values[1] == FloatVal::INFINITY
-		{
-			return FLOAT_SET_INF.clone();
-		}
-
-		Self::new_box(value_storage::Init {
-			ty: ValType::FloatSet,
-			len: (values.len() / 2) as u32,
-			ref_count: 1.into(),
-			weak_count: 0.into(),
-			values: InitEmpty,
-			ints: InitEmpty,
-			floats: FromIterPrefix(values.into_iter()),
-			bytes: InitEmpty,
-		})
-	}
-}
-
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, PartialEq)]
 pub enum DataView<'a> {
 	Int(IntVal),
 	Float(FloatVal),
@@ -922,7 +348,44 @@ pub enum DataView<'a> {
 	Str(&'a str),
 	IntSet(IntSetView<'a>),
 	FloatSet(FloatSetView<'a>),
+	BoolVar(BoolVarRef<'a>),
+	IntVar(IntVarRef<'a>),
+	FloatVar(FloatVarRef<'a>),
+	IntSetVar(IntSetVarRef<'a>),
 }
+
+macro_rules! define_var_ref {
+	($v:ident, $r:ident) => {
+		#[derive(Debug)]
+		pub struct $r<'a> {
+			guard: std::sync::MutexGuard<'a, $v>,
+		}
+
+		impl PartialEq for $r<'_> {
+			fn eq(&self, other: &Self) -> bool {
+				*self.guard == *other.guard
+			}
+		}
+		impl Eq for $r<'_> {}
+		impl std::ops::Deref for $r<'_> {
+			type Target = $v;
+			fn deref(&self) -> &Self::Target {
+				&self.guard
+			}
+		}
+		impl std::ops::DerefMut for $r<'_> {
+			fn deref_mut(&mut self) -> &mut Self::Target {
+				&mut self.guard
+			}
+		}
+		impl<'a> From<std::sync::MutexGuard<'a, $v>> for $r<'a> {
+			fn from(guard: std::sync::MutexGuard<'a, $v>) -> Self {
+				Self { guard }
+			}
+		}
+	};
+}
+pub(crate) use define_var_ref;
 
 #[bitsize(8)]
 #[derive(Clone, PartialEq, Eq, TryFromBits)]
@@ -951,6 +414,7 @@ enum ValType {
 	IntSetVar,
 }
 
+#[allow(dead_code)] // attributes accessed through [`.refs()`]
 #[define_varlen]
 struct ValueStorage {
 	/// Type of the value
@@ -968,14 +432,12 @@ struct ValueStorage {
 	// Number of weak references (e.g., CSE)
 	weak_count: AtomicU8,
 
-	#[allow(dead_code)] // accessed through [`.refs()`]
 	#[varlen_array]
 	values: [Value; match *ty {
 		ValType::Seq => *len as usize,
 		ValType::View => 1,
 		_ => 0,
 	}],
-	#[allow(dead_code)] // accessed through [`.refs()`]
 	#[varlen_array]
 	ints: [i64; match *ty {
 		ValType::Int => 1,
@@ -983,14 +445,32 @@ struct ValueStorage {
 		ValType::View => ViewType::from_len(*len).int_len(),
 		_ => 0,
 	}],
-	#[allow(dead_code)] // accessed through [`.refs()`]
 	#[varlen_array]
 	floats: [FloatVal; match *ty {
 		ValType::Float => 1,
 		ValType::FloatSet => *len as usize * 2,
 		_ => 0,
 	}],
-	#[allow(dead_code)] // accessed through [`.refs()`]
+	#[varlen_array]
+	bool_var: [Mutex<BoolVar>; match *ty {
+		ValType::BoolVar => 1,
+		_ => 0,
+	}],
+	#[varlen_array]
+	int_var: [Mutex<IntVar>; match *ty {
+		ValType::IntVar => 1,
+		_ => 0,
+	}],
+	#[varlen_array]
+	float_var: [Mutex<FloatVar>; match *ty {
+		ValType::IntVar => 1,
+		_ => 0,
+	}],
+	#[varlen_array]
+	int_set_var: [Mutex<IntSetVar>; match *ty {
+		ValType::IntSetVar => 1,
+		_ => 0,
+	}],
 	#[varlen_array]
 	bytes: [u8; match *ty {
 		// Fully stored as bytes
@@ -1044,10 +524,6 @@ unsafe impl<T> ArrayInitializer<T> for InitEmpty {
 
 #[cfg(test)]
 mod tests {
-	use std::iter::empty;
-
-	use expect_test::expect;
-	use itertools::Itertools;
 
 	use super::*;
 
@@ -1073,70 +549,6 @@ mod tests {
 	}
 
 	#[test]
-	fn test_bool_value() {
-		let f: bool = Value::from(false).try_into().unwrap();
-		assert_eq!(f, false);
-		let t: bool = Value::from(true).try_into().unwrap();
-		assert_eq!(t, true);
-	}
-
-	#[test]
-	fn test_integer_value() {
-		let zero: i64 = Value::from(0i64).try_into().unwrap();
-		assert_eq!(zero, 0i64);
-
-		let one: i64 = Value::from(1i64).try_into().unwrap();
-		assert_eq!(one, 1i64);
-		let minus_one: i64 = Value::from(-1i64).try_into().unwrap();
-		assert_eq!(minus_one, -1i64);
-
-		// Unboxed min and max
-		let minimum: i64 = Value::from(Value::MIN_INT).try_into().unwrap();
-		assert_eq!(minimum, Value::MIN_INT);
-		let maximum: i64 = Value::from(Value::MAX_INT).try_into().unwrap();
-		assert_eq!(maximum, Value::MAX_INT);
-
-		// Positive and Negative Infinity
-		let pos_inf: IntVal = Value::from(IntVal::InfPos).try_into().unwrap();
-		assert_eq!(pos_inf, IntVal::InfPos);
-		let neg_inf: IntVal = Value::from(IntVal::InfNeg).try_into().unwrap();
-		assert_eq!(neg_inf, IntVal::InfNeg);
-
-		// i64 min and max
-		let minimum: i64 = Value::from(i64::MAX).try_into().unwrap();
-		assert_eq!(minimum, i64::MAX);
-		let maximum: i64 = Value::from(i64::MIN).try_into().unwrap();
-		assert_eq!(maximum, i64::MIN);
-	}
-
-	#[test]
-	fn test_float_value() {
-		let zero: f64 = Value::from(0.0f64).try_into().unwrap();
-		assert_eq!(zero, 0.0);
-		let one: f64 = Value::from(1.0f64).try_into().unwrap();
-		assert_eq!(one, 1.0);
-		let minus_one: f64 = Value::from(-1.0f64).try_into().unwrap();
-		assert_eq!(minus_one, -1.0);
-
-		let twodottwo: f64 = Value::from(2.2f64).try_into().unwrap();
-		assert_eq!(twodottwo, 2.2);
-
-		// Positive and Negative Infinity
-		let pos_inf: f64 = Value::from(f64::INFINITY).try_into().unwrap();
-		assert_eq!(pos_inf, f64::INFINITY);
-		let neg_inf: f64 = Value::from(f64::NEG_INFINITY).try_into().unwrap();
-		assert_eq!(neg_inf, f64::NEG_INFINITY);
-		// f64 min and max
-		let minimum: f64 = Value::from(f64::MAX).try_into().unwrap();
-		assert_eq!(minimum, f64::MAX);
-		let maximum: f64 = Value::from(f64::MIN).try_into().unwrap();
-		assert_eq!(maximum, f64::MIN);
-
-		assert_eq!(FLOAT_INF_NEG.ref_ty(), RefType::Boxed);
-		assert_eq!(FLOAT_INF_POS.ref_ty(), RefType::Boxed);
-	}
-
-	#[test]
 	fn test_string_value() {
 		let empty = Value::from("");
 		assert_eq!(empty.deref(), DataView::Str(""));
@@ -1154,162 +566,5 @@ mod tests {
 		Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."#;
 		let vlorem = Value::from(lorem);
 		assert_eq!(vlorem.deref(), DataView::Str(lorem));
-	}
-
-	#[test]
-	fn test_sequence() {
-		let empty: Value = empty::<Value>().collect();
-		assert_eq!(empty.deref(), DataView::Seq(SeqView::Direct(&[])));
-		assert!(empty.is_constant(&EMPTY_SEQ));
-
-		let single: Value = [Value::from(1)].into_iter().collect();
-		assert_eq!(single.deref(), DataView::Seq(SeqView::Direct(&[1.into()])));
-
-		let tup2: Value = [Value::from(1), 2.2.into()].into_iter().collect();
-		assert_eq!(
-			tup2.deref(),
-			DataView::Seq(SeqView::Direct(&[1.into(), 2.2.into()]))
-		);
-
-		let list = (1..=2000).map(Value::from).collect_vec();
-		let vlist: Value = list.iter().cloned().collect();
-		assert_eq!(vlist.deref(), DataView::Seq(SeqView::Direct(&list)));
-
-		// Test reverse
-		let blist = vlist.transpose([-1]);
-		let DataView::Seq(view) = blist.deref() else {
-			unreachable!()
-		};
-		assert!(itertools::equal(view.iter(), list.iter().rev()));
-
-		let list = (1..=100).map(|_| Value::from(1i64)).collect_vec();
-		let vlist: Value = list.iter().cloned().collect();
-		let DataView::Seq(view) = vlist.deref() else {
-			unreachable!()
-		};
-
-		assert!(matches!(
-			view,
-			SeqView::Compressed {
-				dims: _,
-				repeat: 99,
-				storage: _
-			}
-		));
-		assert_eq!(view.len(), list.len());
-		itertools::equal(list.iter(), view.iter());
-
-		let vlist: Value = (1..=9).map(Value::from).collect();
-		// Test dimensions
-		let dim_list = vlist.with_dim([(-1, 1), (5, 7)]);
-		let DataView::Seq(view) = dim_list.deref() else {
-			unreachable!()
-		};
-		for i in -1..=1 {
-			for j in 5..=7 {
-				assert_eq!(view[&[i, j]], ((i + 1) * 3 + j - 4).into());
-			}
-		}
-		// Test slicing
-		for i in 1..=3 {
-			// Slice row
-			let row = dim_list.slice([(i - 2, i - 2), (5, 7)], [(0, 2)]);
-			let DataView::Seq(view) = row.deref() else {
-				unreachable!()
-			};
-			assert_eq!(view.len(), 3);
-			assert_eq!(view.iter().count(), 3);
-			itertools::equal(
-				view.iter().cloned(),
-				(1..=3).map(|j| Value::from((i - 1) * 3 + j)),
-			);
-			// Slice column
-			let row = dim_list.slice([(-1, 1), (i + 4, i + 4)], [(1, 3)]);
-			let DataView::Seq(view) = row.deref() else {
-				unreachable!()
-			};
-			assert_eq!(view.len(), 3);
-			assert_eq!(view.iter().count(), 3);
-			itertools::equal(
-				view.iter().cloned(),
-				(1..=3).map(|j| Value::from((j - 1) * 3 + i)),
-			);
-		}
-		// Test transpose
-		let t = dim_list.with_dim([(1, 3), (1, 3)]).transpose([2, 1]);
-		let DataView::Seq(view) = t.deref() else {
-			unreachable!()
-		};
-		for i in 1..=3 {
-			for j in 1..=3 {
-				assert_eq!(view[&[i, j]], ((j - 1) * 3 + i).into());
-			}
-		}
-	}
-
-	#[test]
-	fn test_set() {
-		let isv_empty = INT_SET_EMPTY.clone();
-		let DataView::IntSet(sv) = isv_empty.deref() else {
-			unreachable!()
-		};
-		expect!["∅"].assert_eq(&sv.to_string());
-
-		let isv_inf = INT_SET_INF.clone();
-		let DataView::IntSet(sv) = isv_inf.deref() else {
-			unreachable!()
-		};
-		expect!["int"].assert_eq(&sv.to_string());
-
-		let isv_simple = Value::from_iter([IntVal::Int(-3)..=5.into()]);
-		let DataView::IntSet(sv) = isv_simple.deref() else {
-			unreachable!()
-		};
-		expect!["-3..5"].assert_eq(&sv.to_string());
-		assert!(itertools::equal(sv.values(), (-3..=5).map_into()));
-
-		let isv_open_left = Value::from_iter([IntVal::InfNeg..=5.into()]);
-		let DataView::IntSet(sv) = isv_open_left.deref() else {
-			unreachable!()
-		};
-		expect!["..5"].assert_eq(&sv.to_string());
-
-		let isv_open_right = Value::from_iter([0.into()..=IntVal::InfPos]);
-		let DataView::IntSet(sv) = isv_open_right.deref() else {
-			unreachable!()
-		};
-		expect!["0.."].assert_eq(&sv.to_string());
-
-		let isv_gaps = Value::from_iter([
-			IntVal::Int(-3)..=(-3).into(),
-			0.into()..=0.into(),
-			3.into()..=5.into(),
-		]);
-		let DataView::IntSet(sv) = isv_gaps.deref() else {
-			unreachable!()
-		};
-		expect!["-3..-3 ∪ 0..0 ∪ 3..5"].assert_eq(&sv.to_string());
-		assert!(itertools::equal(
-			sv.values(),
-			[-3i64, 0, 3, 4, 5].into_iter().map_into()
-		));
-
-		let fsv_empty = FLOAT_SET_EMPTY.clone();
-		let DataView::FloatSet(sv) = fsv_empty.deref() else {
-			unreachable!()
-		};
-		expect!["∅"].assert_eq(&sv.to_string());
-
-		let fsv_inf = FLOAT_SET_INF.clone();
-		let DataView::FloatSet(sv) = fsv_inf.deref() else {
-			unreachable!()
-		};
-		expect!["float"].assert_eq(&sv.to_string());
-
-		let fsv_simple = Value::from_iter([FloatVal::from(-2.3)..=4.3.into()]);
-		let DataView::FloatSet(sv) = fsv_simple.deref() else {
-			unreachable!()
-		};
-		expect!["-2.3..4.3"].assert_eq(&sv.to_string());
 	}
 }
